Backend/requirements.txt
click==7.1.1
dnspython==1.16.0
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mock==4.0.2
pymongo==3.10.1
Werkzeug==1.0.0
